move build/request local stellarsvc->stellar
